+++ src/http/websocket/index.ts
@@ -11,7 +11,12 @@ export type { WebsocketRouteOptionsType } from './WebsocketRouteOptions.js';
 export { WebsocketServerActor } from './WebsocketServerActor.js';
 
 export { WebsocketClientActor } from './WebsocketClientActor.js';
-export { WebsocketClientOptions, WebsocketClientOptionsBuilder, WebsocketClientOptionsValidator } from './WebsocketClientOptions.js';
+export {
+  DEFAULT_WEBSOCKET_INBOUND_LOW_WATER_MARK_DIVISOR,
+  WebsocketClientOptions,
+  WebsocketClientOptionsBuilder,
+  WebsocketClientOptionsValidator,
+} from './WebsocketClientOptions.js';
 export type { WebsocketClientOptionsType } from './WebsocketClientOptions.js';
 
 export type { WebsocketConnection } from './WebsocketConnection.js';